Eliminate many cases of dynamic class loading by generating classloading methods
Previous for annotation metadata references to classes were stored as Strings and then Class.forName(..)
was used to load the classes. This caused isses on GraalVM. This change adds static methods to generated
classes that perform class loading for annotation metadata without using Class.forName(..) eliminating a lot
of dynamic classloading in a Micronaut application. Fixes #677