Using `ConsulConfiguration` subclass instead of the parent one.
When using Consul discovery in a cloud environment, there can be multiple
instances of a bean with type `DiscoveryClientConfiguration` (eg:
`Route53ClientDiscoveryConfiguration` and `ConsulConfiguration`. In this
scenario, Micronaut will fail to build the application context because
it won't be able to wire a `DiscoveryClientConfiguration` candidate in
`ConsulServiceInstanceList` when there's more than one bean of that type.