perf/cleanup: redundant if/else on return (#1168)
* Optimize toArray (#1157)

replace toArray(new T[size]) with toArray(new T[0]) for better performance

* perf/cleanup: redundant if/else on return

Remove redundant else / else if if a return is present. Saves a bytecode instruction and cleans up the code slightly for better reading. Note also that HashMap's max size is size * 0.75, so it needs to be sized at size / 0.75.

* fix incorrect elseif from previous change