Fix incorrect usage of Androids onTerminate() in doc
A concept that is foreign for non-Android devs, is that there are a bunch of entry 
lifecycle points but no exit lifecycle for the whole _application_ (there is for `Activity`
ie. the Android OS interface for creating views). So there is no concept of "exiting application",
it just disappears from the heap at some point, the Android OS handles this. As developer you should
therefore not rely on cleanup code in the `Application`, only in `Activities` (or similar) - which are also not
guaranteed to be called (specifically `onDestroy()`).

In this particular case, this is even worse because `onTerminate()` is _only_ called in emulator
environments (check out: https://developer.android.com/reference/android/app/Application.html#onTerminate())
which makes this example highly misleading. As the cleanup code is probably not useful to execute when Android
OS decides to clear the app from the heap, it is redundant anyway.