Feature: Native Options in Netty
This changeset adds the ability to set options supported by native
Netty-supported transports, like `epoll` and `kqueue`.

Micronaut already makes it easy to specify `options` and `childOptions`
under Netty, but those options don't offer support for higher-level
stuff exclusively supported by transports like `epoll` and `kqueue`.

- Users in certain circumstances may want to turn on `TCP_NODELAY`.
  This is already easy via Micronaut's support for Netty options,
  because `TCP_NODELAY` is part of `io.netty.channel.ChannelOption`,
  which is where Micronaut looks for socket options.

- Users on Linux might want to take John Nagle's advice and, instead of
  enabling `TCP_NODELAY`, enabling the more modern `TCP_QUICKACK`
  option (supported since Linux 2.4.4). This can't be done out of the
  box in Micronaut, however, because that option is defined in Netty at
  `io.netty.channel.epoll.EpollChannelOption`.

This fixes and closes #2066, which I filed to track the issue. Basically,
`ChildOption` instances in the `NettyHttpServerConfiguration` are now
wrapped in an object that describes whether they are base options, or
options for specific transports.

This enables a few new config options that look like this:
```yaml
micronaut:
  server:
    netty:
      childOptions:
        soKeepalive: true
      epoll:
        childOptions:
          tcpQuickack: true
      kqueue:
        childOptions:
          tcpNopush: true
```

At runtime, the transport is detected if `use-native-transport` is turned
on. If it's `epoll` or `kqueue`, their individual options are merged into
a combined map, with the newly-wrapped `ChildOption` instances. This
enables `processOptions` to resolve the option to set on the appropriate
class.

- [x] Add `NettyChannelOption` private wrapper class in
      `NettyHttpServerConfiguration`
- [x] When gathering options, wrap them in `NettyChannelOption`
- [x] When applying options, resolve them via `NettyChannelOption`
- [x] Declare new configuration options
- [ ] Add unit testing
- [ ] Testing downstream in production