Dual port patch
After getting this into real use in the micronaut-acme project found
that it was firing the startup event twice now. This is not ideal as it
triggered the certificate validation process twice.

Also it looks like if using a random port failed to bind and it retried
it would have also suffered from this
(Not sure much would have been affected by this other than tests cases)