Fix the rest of the compile issues.
./gradlew clean testClasses now runs successfully.

Note : Had to explicity add junit-api jar as a dependency in 2 of the
test-suite projects because i believe switching to gradle 5 caused
resolution of that dep to now not be the same.

Micronaut-test project will also need to be tweak to get that to work
and not have to duplicate it here is what the current thought is.