Allow introduction advice to invoke an alternative object
Currently invoking the executableMethod reference in a method interceptor in introduction advice
allows results in an AbstractMethodError. This prevents the ability to invoke an alternative implementation
when introducing new behaviour. This commit alters that for newly compiled introduction advice allowing
an alternative implementation to be invoked, effectively supporting a delegation pattern.