Remove use of computeIfAbsent. (#3632)
This change removes the use of computeIfAbsent which for some cases leads to recursive update.
There may be a slight cost in terms of duplicate initialization of BeanDefinition caches dependency on the injection path,
but I don't think it is worth the cost of tring to retain atomicity of computeIfAbsent since these are just the caches
for the backing bean definitions and not the bean instances themselves. We may even want to attempt a periodic purge
of these caches to save memory, but that is a separate issue.

I have not been able to come up with a test that replicates the issue, but have verified it fixes in the problem
by testing in the Micronaut Security code base. Once Security upgrades to 2.0.1 the intermittent failure should be fixed there.

Fixes #3577