A clearer expression about @Retryable
Hi there. Currently, @Retryable annotation works only when the function throws RuntimeException or its subclass.
But for instance, IOException and other significant exceptions don't belong to it. So we have to wrap it with RuntimeException then rethrow.
Can't discuss whether it's what it should be or not. But I thought at least the document could mention about it to avoid confusion and having to look at the source code.
Thanks.