Fix dependency resolution happening at configuration time
The BOM check tasks were not using `doLast`, which means that
everything was done at configuration time, which is particularly
bad in this case because it resolves _all_ configurations, and
also requires _all_ artifacts to be downloaded, for every build.

This is a major performance issue, which this commit just fixes
by wrapping everything into an appropriate `doLast`.

Ideally this code should be moved in a `buildSrc` plugin, in order
to avoid bloating the build script, and which would have made
this evident (because the code would have lived in a `@TaskAction`).