Issue 4244 - ensure that the environments list is mutable (#34244)
Since the BootstrapEnvironment calls the contructor with an implementation of `getEnvironments`
that is immutable, as could others, the DefaultEnvironment class was updated to convert the list
to a mutable implementation.

The test case for getting Micronauts environments from the environment variables has been updated
to reflect the case of setting the bootstrap environment when the environment variable is used.