Get name from PackageElement.getQualifiedName().
The existing code assumes that the toString() of the TypeElement
of the package will return the package name. This is not true
for the eclipse ecj compiler, where the toString() will return
"package " + the package name. For example, a TypeElement
representing io.micronaut would return "package io.micronaut"
on ecj.

This change will get the qualified name from the type element if
it is a PackageElement instance, then use toString() on the
Name that is returned. If the element is not a PackageElement,
fallback to the toString() directly on the type element.