Reworks the CustomScope interface to support disposing dependent instances.
Currently if a bean that is injected into another instance doesn't have a defined scope and it defines a PreDestroy method what happens is that when the parent bean is destroyed then this method is never called.

This PR alters the behaviour so that dependent instances are tracked and associated with the BeanRegistration and then when the parent bean is installed any child beans are also destroyed.

In addition to make it possible to write custom scopes that allow disposable of downstream dependent beans the CustomScope interface has been reworked introducing two new interfaces called BeanCreationContext and CreatedBean. The latter of these implements AutoCloseable such that it can be used to dispose of any downstream dependent beans.

The result of this is the removal of a lot of custom bean destruction code that existed in some of the scopes like RequestScope and RefreshableScope.