Override status(HttpStatus) instead of status(HttpStatus, CharSequence)
Refactor override of MutableHttpResponse's method:
- status(HttpStatus, CharSequence), to
- status(HttpStatus)

This override is done by MutableHttpResponse implementations.

HttpStatus couldn't hold a custom reason phrase, since now it can then
the reason is set in an enum value, so every HttpStatus value has a reason.
This makes an enum value the source of truth for both the code and the reason.