Remove support for runtime classpath scanning with ASM
This change removes the support for runtime classpath scanning with ASM. This functionality is only used by the Hibernate/JPA module and in all recent versions of Micronaut 2.x entities are located via introspections and not via this API so it is no longer necessary.

Keeping this around significantly increases the native image size for Micronaut since it makes ASM reachable and pulls in a large amount of unnecessary code at runtime.