Introduce a version catalog
This commit replaces the "dependencyVersion" and "dependencyModuleVersion"
calls with Gradle's version catalogs. In the process, I tried to clarify
what belongs to the BOM (as "managed" dependencies") vs what is not.

Currently this breaks BOM verification module, this will be fixed in
another commit. It will also probably break the automatic upgrading of
dependencies. However, the result is much cleaner: there are no duplicate
coordinates anymore in build files.