Publish a version catalog alongside the BOM
This commit introduces a new artifact published alongside the BOM:
a Gradle _version catalog_. A version catalog is not a replacement
for a BOM. It can be seen as an alternative, or a complement to a
BOM file.

For example, a user can import the Micronaut version catalog in
their settings file by doing:

```gradle
dependencyResolutionManagement {
    versionCatalogs {
        create("mn") {
            from("io.micronaut:micronaut-bom:3.0.1-SNAPSHOT")
        }
    }
}
```

Gradle will then generate _version accessors_, which allows
replacing dependency notations from:

```gradle
implementation "ch.qos.logback:logback-classic"
```

to:

```gradle
implementation mn.logback
```

Those accessors are _type safe_. Version catalogs should
really be seen as "recommendations". Without applying a
BOM in addition, the versions declared in a catalog have
_no impact_ on dependency resolution. Therefore, it is
often recommended to apply both the BOM _and_ use catalogs
to declare dependencies.

It comes with an advantage, which is that versions declared
in the BOM can be overridden by the user:

```gradle
dependencyResolutionManagement {
    versionCatalogs {
        create("mn") {
            from("io.micronaut:micronaut-bom:3.0.1-SNAPSHOT")
            version("snakeyaml") {
                strictly("1.28")
            }
        }
    }
}
```