Add a sanity check to make sure that the catalog is valid
In case a managed library references a non managed version, the build
used to fail with a weird message (either string index out of bounds
error, or a truncated version name).

The build will now fail with a proper message explaining that you
have to use a managed version too.