Introduce static optimizations registry
This is preliminary work for Micronaut AOT, a module which is
going to compute optimizations at build time. For this to work,
we need to introduce a class called `StaticOptimizations` which
acts as a registry for "static" optimizations. The registry
works by statically (in the sense of a static field) injecting
optimization data, that _specific_ classes care about.

The pattern is simple: if a class can benefit from precomputed
data at build time, then it should create an inner class which
is going to store this data as an "optimization". The AOT module
will eventually be responsible for computing the data and injecting
it. This means that by themselves, the changes in this PR do nothing:
the optimizations need to be injected at runtime.

As an example, let's imagine that we have a class called `Primes` which
computes prime numbers. An obvious optimization that we can do is to
precompute a number of primes at build time, and inject them to the
`Primes` class. Therefore, the pseudo-code would look like this:

```
public class Primes {
    public static Optimizations {
       private final List<Integer> firstPrimes;
       ...
    }

    private List<Integer> primes = StaticOptimizations.get(Optimizations)
        .map(opt -> new ArrayList<>(opt.firstPrimes)
        .orElse(new ArrayList<>);
}
```

The pattern is to pre-initialize the fields by looking if a static optimization
is available.

The Micronaut AOT module will then be responsible for _somehow_ calling the
`StaticOptimizations#set` method at runtime, typically when the application
starts.

This mechanism is interesting for different reasons:

- it makes all static optimizations discoverable by searching for usages of
the `StaticOptimizations.get()` method
- it's transparent, in the sense that if no optimizations are injected at
runtime, the behavior of the application is rigourously the same as before
- it doesn't matter how the optimizations are injected. Micronaut AOT is one
way, but nothing prevents user code from injecting optimizations themselves
- it's runtime independent: distinct optimizations can be injected for
different contexts. For example, we can inject optimizations specifically
for the JIT mode, and different optimizations for native images. This is
typically the case in this commit for the service loading mechanism.