Introduce sanity check for generated services
This commit introduces an _internal_ sanity check for generated
services. It happens that during the build, annotation processing
fails and doesn't generate the expected number of services. When
this happens it's pretty hard to figure out, and you can realize
late.

To workaround this, we introduce an extension which asks for the
number of expected services. If after compilation there's a mismatch
the build will fail (this can be disabled).