Add new `@AccessorsStyle` annotation
This annotation is used to explicitly tell Micronaut that the getters
and setters for a Java Bean are not named in a standard way, i.e., using
get/set prefix.
At this moment we only do something for NONE style which is the one used
by Lombok "fluent" accessors.

See https://github.com/micronaut-projects/micronaut-openapi/issues/593