Support WebSocket permessage-deflate extension (#6487)
* Support WebSocket permessage-deflate extension
Both for client and server.
On the server side, this requires a change in the pipeline setup logic. Because the permessage-deflate codecs are placed after the websocket codecs, but aren't always present, we can't simply place the `NettyServerWebSocketHandler` after the `wsdecoder` anymore. Instead, we place it at the old location of the upgrade handler.

Resolves #6069

* slice websocket message ByteBuf to avoid changing indices

* properly forward non-http messages in HttpRequestCertificateHandler