Introduce an internal interface for finalizing context configuration
During debugging of Micronaut AOT, we discovered that some bean requirements
weren't met, because bean configurations were not read. This happened because
those beans are only read when the context is actually started. In AOT, we
don't want to start the context, only introspect it, so we missed an operation
to perform the finalization of configuration, while keeping the context in
an unstarted state.

For this purpose, we introduce a new internal interface, which AOT will be
able to check for, and call this method instead of using reflection to
call the `readAllXXX` methods.