Introdce a static optimizations loader (#6694)
This commit introduces loading of "static optimizations" via service loading.
The reason we have to do this is, again, _ordering_. When implementing AOT
optimizations for lambdas, we realized that it was possible for some classes
on classpath to call the `SoftServiceLoader` in a static initializer, _before_
the static optimizations were set. The consequence of doing this is that once
the AOT optimizer has optimized service loading, then no service would be
found because when this call is made, the optimizations are not injected yet.

As a consequence, we also introduced a sanity check to make sure that the
static optimizations `get()` method are only called _after_ the `set()`
method was called. It is illegal to set an optimization after the first
consumer has read its state.

Ideally, we should get rid of static initializers in beans, but there's a
long way to go. This commit therefore introduces a pragmatic solution to
make sure that optimizations are always set before any consumer accesses
them.

This will also require a change in Micronaut AOT to use this new API,
instead of injecting optimizations via a static block in a generated
context configurer.