Allow file watcher restarts for messaging apps
The file watcher required an EmbeddedServer to be enabled.

If you create a Kafka app with no http enabled, ie:

mn create-messaging-app messaging-test --features kafka

then you get a MessagingApplication in place of an EmbeddedServer.

This change moves the requirement up to an EmbeddedApplication<?> instead of a specific
EmbeddedServer implementation of it, so we listen for restart events for that type.