Don't pass Host header by default in ProxyHttpClient
Currently, the proxy method passes through the `Host` header from the originating service request, which is incorrect.

The proxy method should rewrite the `Host` header to the `host()` value of the URI being called by `proxy()`.

This change defaults to replacing the `Host` header with the proxied host, and adds options to `KEEP` the old method,
copy the `Host` to `X-Forwarded-Host`, or append it to `Forwarded`.