Fix record serialization for native image Without `@ReflectiveAccess` on native-image, jackson fails serialization and deserialization of record types with an exception message `Failed to access RecordComponents of type ...` (see issue). This happens before the `BeanIntrospectionModule` comes into play, so mapping is not possible even though the `BeanIntrospectionModule` could have replaced the missing reflection information.
This patch hooks the jackson `AccessorNamingStrategy` to bypass the relevant code for serialization. This seems reasonable enough. It also hooks `BeanDeserializerFactory` for deserialization, but both the approach in patching `_constructDefaultValueInstantiator` and the way the factory is inserted (can't be done in the `BeanIntrospectionModule`) are very hacky.

I hope to fix this upstream in jackson at some point, once I get CLA approval, but we're still on 2.12 so we couldn't take advantage of that for a while still.

Fixes #6277