Move server SSL config to the right place
The config items for server SSL configuration were partially in the Server config
and partially in the root micronaut.ssl.* config.

This change moves:

- enabled
- port
- build-self-signed

into micronaut.server.ssl.*

`port` and `build-self-signed` are deprecated in the common config, and setting
them will emit a warning log line to inform the user of their imminent removal.

Once they are removed, we can get rid of the custom conditions on both `SelfSignedBuilder`
and `CertificateProvidedSslBuilder` and instead simply require the correct properties