Add an ApplicationContextBuilder method to fail fast on invalid config
previously if you had a list of values (especially enums) and one or more of them
were invalid, they were silently dropped, and the app started as per normal.

This change adds an ApplicationContextBuilder method failOnConfigConversionError which
if set to true when the app starts up will thrown an exception if one of these conversion
errors is seen.