Add unsupported config option for logging HTTP data to pcap file This patch adds an option to the netty config to set a file path pattern where pcap dumps are written to. For every connection, two files are created, one for the "full" stream (qualifier: 'encapsulated'), and one with the ssl removed (qualifier: 'ssl-decapsulated'). I've been using similar code for debugging HTTP issues for a while, especially the decapsulated approach. Adding this to core behind a config option allows us to use this logging for hard-to-reproduce issues as well.
Because the config option is marked as `@Internal`, and the code path does nothing if the option is not set, there should be little risk to this change.
To prevent users shooting themselves in the foot by forgetting to turn off this option, there is a log warning on every request.