Do not wait for request data after decode failure (#6941)
When there is a failure during decoding of header values, netty will consume the entire input and not forward it. However, if we expect content data e.g. due to a Content-Length header, RoutingInBoundHandler would wait indefinitely for that data before sending the error response.

This patch changes HttpStreamsServerHandler to treat a decode failure request as if it had no body. This way a StreamedHttpRequest is never created. I also considered just bypassing the StreamedHttpRequest wait in RoutingInBoundHandler, but that risks unexpected behavior from the StreamedHttpRequest publisher never finishing (see #6935), so I tried to avoid StreamedHttpRequest entirely.

There is also a change to force the connection to close, to prevent possible downstream errors from connection reuse.

Fixes #6925