Do not request demand for completed HandlerPublisher (#6949)
* Do not request demand for completed HandlerPublisher
When a HandlerPublisher is removed, it is completed. If at that point it is in DEMANDING state, this will trigger a `flushBuffer()`, which in turn can trigger a `requestDemand()` and a `context.read()`. This read will happen before the removal of the HandlerPublisher has finished, which can lead to a race condition upstream in HttpStreamsHandler: currentlyStreamedMessage will be set to null *after* the removal of the publisher, at which point a new publisher has already been created from the `context.read()`.

This patch fixes this issue in two ways: the HandlerPublisher won't trigger a read anymore if it's already completed, and the HttpStreamsHandler will unset the currentlyStreamedMessage *before* triggering downstream handler removal.

* fix flaky test

* Revert "fix flaky test"

This reverts commit 2eb55e0dd8cd73814badc7bcf571e5dca9b79f71.

* clean up ServerRequestContext in embedded test