bug: Don't strip newlines proxying ServerSentEvents
When the HttpClient.proxy is used to proxy an SSE stream, we were stripping
delimiters from the data, which caused consumers of this that were sending;

```
Accept: text/event-stream
```

To get a single stripped event at the end of the stream, instead of the events
arriving individually as expected.

This change passes that we are proxying down to the `HttpClientInitializer`
so we can stop stripping delimiters if this is the case.

Fixes: #6985

Huge thanks to @yawkat for helping triage this