Refactor the Gradle build
This commit refactors the Gradle build to use the new Micronaut
build plugins and move it to the new build template (which makes
use of convention plugins).

The build is obviously more complex than for external Micronaut
modules, in particular because some modules are providing core
functionalities but also depend on external Micronaut modules
for tests (e.g Micronaut Test).