Update to jackson 2.13
Jackson minor versions are not 100% compatible, so we held off on updating before this. However we have to move up eventually because 2.12 isn't well-supported anymore. Now, a HIGH severity CVE has been filed for a jackson issue ( https://github.com/FasterXML/jackson-databind/issues/2816 , though the actual impact is debatable). While the patch for the CVE isn't available yet, we need to move to jackson 2.13.2 for micronaut 3.4 so that we can do the update to 2.13.3 in a micronaut patch release.