Add grpc back in as a managed version
For 3.5.0, we added a BOM to GRPC and used it here.  The grpc version is required in the starter
module however, and removing that caused it to break.

I tried and failed to fix it here https://github.com/micronaut-projects/micronaut-starter/pull/1162

This change adds back in the GRPC bom, which was previously removed in

https://github.com/micronaut-projects/micronaut-core/pull/7162

And updates the grpc and protobuf versions to align with the current versions in the grpc
module v3.2.0