Use ServiceCollector for SoftServiceLoader.iterator The new directory-based service discovery introduced by #7195 was implemented for ServiceCollector, but not for the old SoftServiceLoader.iterator method.
On a high level, this patch makes the iterator use the same collector logic as collectAll. Unfortunately, this revealed an issue with the fork-join approach: the service loader is often used in a static initializer, but multithreading is problematic there (can lead to class initialization deadlocks).

To avoid this issue, I made the multithreading opt-out, and refactored the RecursiveActions to facilitate optional single-threaded operation.

It may be better to make the multithreading opt-in instead, since using the service loader in a static initializer seems to be the norm, not the exception.

I also refactored various use sites of the SoftServiceLoader to use higher-level methods like collectAll and first.

Fixes #7594