Clean up request even when using a customizable invoker (#7728)
When the response body is a `NettyCustomizableResponseTypeHandlerInvoker` (e.g. from an InputStream), `RoutingInBoundHandler` would never clean up the corresponding request.

This patch makes the `NettyCustomizableResponseType.write` and related classes return the `ChannelFuture` for the last response write operation they trigger. `RoutingInBoundHandler` then listens to that future and cleans up the associated request when the write completes.

Fixes #7704