Use v6.0.0 of the internal build tools
I had to work around a bug in Gradle and open the JDK package for inject-java-test, but I think this is just used in tests.

The plugin pulls in io.micronaut:micronaut-inject-groovy which can't be resolved versionless.

So I added a resolution to switch this with the :inject-groovy module, which I think is more sensible here