Split the runtime module into multiple modules (#8235)
This PR splits the `micronaut-runtime` module into multiple modules.  The new modules are:

* `micronaut-retry` - The retry logic implementation is now optional. In order to support this by default declarative http clients are no longer automatically `@Recoverable` (Fixes #3719). This is a better arrangement as not everyone needs or uses fallbacks and recovery
* `micronaut-discovery-core` - The service discovery implementation is split into a separate module that is no longer required. The HTTP client module still depends on it though and would be nice to remove this coupling.

Following this change `micronaut-runtime` is just a meta module that depends on other modules. Other Micronaut modules should update and depend specifically on an individual module such not to pull in more than is necessary.

The motivation for this is to reduce the footprint of a new app since many features are not used (heartbeat, health checks, service discovery, retry etc.) unless you actually need them.

In addition the `micronaut-http-server-netty` module no longer depends on `micronaut-discovery`, `micronaut-health` and `micronaut-websocket`. If users want any of those features then they can add them individually to their application. Micronaut Launch will need to be updated with new feature for WebSocket support. It is probably not necessary to include new features for discovery and health as these will be transitively pulled by the Management module.

Also fixes #8099 by introducing a new interface to resolve the `JsonMapper`

Note this PR also moves some functionality into other modules:

https://github.com/micronaut-projects/micronaut-kotlin/pull/440
https://github.com/micronaut-projects/micronaut-reactor/pull/237

This change also switches Micronaut's integration to GraalVM to be largely runtime initialized including logging. Only Micronaut metadata is build time initialized. This change improves startup time and reduces the size of the images we build.

Co-authored-by: Sergio del Amo <sergio.delamo@softamo.com>