Remove dependencies which are not directly used
This commit removes from the core BOM all managed dependencies which
are not directly used in build scripts. This basically means that
those dependencies are only present so that precisely we can have them
in the BOM. Now that the BOM is extracted in its own project, they are
redundant.