Remove managed libraries which are not used
This commit moves a number of libraries from the managed scope to unmanaged.
The only dependencies which are managed are now the ones which are in api
or implementation scope.

Those dependencies were previously in the BOM, but won't be directly managed
by core. This means a couple of things:

1. the new BOM will have to duplicate those dependencies, in order to manage
them
2. the dependencies will have to be copied in the new BOM (because they are
not present in the core BOM anymore, they won't be inlined like the other
managed dependencies)